requirements.txt
streamlit
opencv-python-headless
tensorflow
numpy
pillow
requests
